+++ cpp/ui/MxcAnimatedImage.cpp
@@ -10,11 +10,10 @@
 #include <QMimeDatabase>
 #include <QQuickWindow>
 #include <QSGImageNode>
-#include <QStandardPaths>
-
 #include <matrix-client-library/EventAccessors.h>
 #include <matrix-client-library/Logging.h>
 #include <matrix-client-library/MatrixClient.h>
+#include "../GlobalObject.h"
 
 void
 MxcAnimatedImage::startDownload()
@@ -54,8 +53,8 @@ MxcAnimatedImage::startDownload()
     const auto url  = mxcUrl.toStdString();
     const auto name = QString(mxcUrl).remove(QStringLiteral("mxc://"));
     QFileInfo filename(
-      QStringLiteral("%1/media_cache/media/%2.%3")
-        .arg(QStandardPaths::writableLocation(QStandardPaths::CacheLocation), name, suffix));
+      QStringLiteral("%1/media/%2.%3")
+        .arg(GlobalObject::instance()->mediaCachePath(), name, suffix));
     if (QDir::cleanPath(name) != name) {
         nhlog::net()->warn("mxcUrl '{}' is not safe, not downloading file", url);
         return;