+++ cpp/ui/MxcMediaProxy.h
@@ -5,33 +5,25 @@
 
 #pragma once
 
-#include <QAbstractVideoSurface>
 #include <QBuffer>
-#include <QMediaContent>
-#include <QMediaPlayer>
 #include <QObject>
 #include <QPointer>
 #include <QString>
+#include <QFileInfo>
 #include "../TimelineModel.h"
 
 class TimelineModel;
 
-// I failed to get my own buffer into the MediaPlayer in qml, so just make our own. For that we just
-// need the videoSurface property, so that part is really easy!
-class MxcMediaProxy : public QMediaPlayer
+class MxcMediaProxy : public QObject
 {
     Q_OBJECT
     Q_PROPERTY(TimelineModel *roomm READ room WRITE setRoom NOTIFY roomChanged REQUIRED)
     Q_PROPERTY(QString eventId READ eventId WRITE setEventId NOTIFY eventIdChanged)
-    Q_PROPERTY(QAbstractVideoSurface *videoSurface READ getVideoSurface WRITE setVideoSurface NOTIFY
-                 videoSurfaceChanged)
-    Q_PROPERTY(bool loaded READ loaded NOTIFY loadedChanged)
-    Q_PROPERTY(int orientation READ orientation NOTIFY orientationChanged)
+    Q_PROPERTY(QUrl mediaFile READ mediaFile NOTIFY mediaFilehanged)
 
 public:
     MxcMediaProxy(QObject *parent = nullptr);
 
-    bool loaded() const { return buffer.size() > 0; }
     QString eventId() const { return eventId_; }
     TimelineModel *room() const { return room_; }
     void setEventId(QString newEventId)
@@ -44,27 +36,21 @@ public:
         room_ = room;
         emit roomChanged();
     }
-    void setVideoSurface(QAbstractVideoSurface *surface);
-    QAbstractVideoSurface *getVideoSurface();
-
-    int orientation() const;
+    QUrl mediaFile() const {return mediaFile_;}
+    void setMediaFile(const QFileInfo &fileinfo);
 
 signals:
     void roomChanged();
     void eventIdChanged();
     void loadedChanged();
-    void newBuffer(QMediaContent, QIODevice *buf);
+    void mediaFilehanged();
 
-    void orientationChanged();
-    void videoSurfaceChanged();
-
-private slots:
-    void startDownload();
+public slots:
+    void checkMediaFileExist();
+    void startDownload(bool justCache);
 
 private:
     TimelineModel *room_ = nullptr;
     QString eventId_;
-    QString filename_;
-    QBuffer buffer;
-    QAbstractVideoSurface *m_surface = nullptr;
+    QUrl mediaFile_;
 };