+++ qml/media/MediaControls.qml
@@ -74,7 +74,7 @@ Rectangle {
         anchors.left: control.left
         anchors.right: control.right
 
-        NhekoSlider {
+        NhekoSlider { 
             Layout.fillWidth: true
             Layout.leftMargin: 4
             Layout.rightMargin: 4
@@ -91,25 +91,31 @@ Rectangle {
             spacing: 4
             Layout.fillWidth: true
 
-            // Cache/Play/pause button
+            // Cache button
+            ImageButton {
+                id: downloadImage
+
+                Layout.alignment: Qt.AlignLeft
+                buttonTextColor: GlobalObject.colors.text
+                Layout.preferredHeight: 20
+                Layout.preferredWidth: 20
+                image: ":/images/download.svg"
+                onClicked: control.loadActivated()
+            }
+            // Play/pause button
             ImageButton {
                 id: playbackStateImage
-
                 Layout.alignment: Qt.AlignLeft
                 buttonTextColor: GlobalObject.colors.text
                 Layout.preferredHeight: 24
                 Layout.preferredWidth: 24
                 image: {
-                    if (control.mediaLoaded) {
-                        if (control.mediaState == MediaPlayer.PlayingState)
-                            return ":/images/pause-symbol.svg";
-                        else
-                            return ":/images/play-sign.svg";
-                    } else {
-                        return ":/images/download.svg";
-                    }
+                    if (control.mediaState == MediaPlayer.PlayingState)
+                        return ":/images/pause-symbol.svg";
+                    else
+                        return ":/images/play-sign.svg";
                 }
-                onClicked: control.mediaLoaded ? control.playPauseActivated() : control.loadActivated()
+                onClicked: control.playPauseActivated()
             }
 
             ImageButton {